+++ glud/matchers.py
@@ -9,7 +9,7 @@ __all__ = [
     'functionDecl', 'has', 'hasName', 'hasReturnType', 'hasStaticStorageDuration',
     'hasTypename', 'isDerivedFrom', 'isSameOrDerivedFrom', 'namespaceDecl',
     'recordDecl', 'stmt', 'typedefDecl', 'unless', 'isDefinition', 'hasAncestor',
-    'isExpansionInFileMatching', 'varDecl', 'hasParent'
+    'isExpansionInFileMatching', 'varDecl', 'hasParent', 'argumentCountIs'
 ]
 
 def allOf(*args):
@@ -534,4 +534,21 @@ def hasParent(*args):
     """
     return ParentMatcher(*args)
 
+
+def argumentCountIs(N):
+    """Matches if a cursor has the specified number of arguments
+
+    >>> from glud import *
+    >>> config = '''
+    ...  int f();
+    ...  int g(int);
+    ...  int h(int, int);
+    ... '''
+    >>> m = functionDecl(argumentCountIs(1))
+    >>> for c in parse_string(config).cursor.walk_preorder():
+    ...     if m(c):
+    ...         print(c.spelling)
+    g
+    """
+    return ArgumentCountMatcher(N)